bot/requirements.txt
aiogram==3.4.1
fastapi==0.114.2
uvicorn[standard]==0.30.6
pdfminer.six==20231228
python-docx==1.1.2
pydantic==2.9.2
python-dotenv==1.0.1
